4月低稳定版 1.增加比赛写字游戏 2.互动修改 3.写字难度修改 4.其他bug修复

+++ Manager/PropsManager.cpp
@@ -16,7 +16,7 @@ static vector<int> aniListToVector(string sAniList){
     string sOrder = sAniList;
     int begin = 0, end = 0;
     while (end != -1) {
-        end = sOrder.find("-", begin);
+        end = (int)sOrder.find("-", begin);
         string tmp = "";
         if (end == -1) {
             tmp = sOrder.substr(begin, sOrder.length() - begin);
